Update manifest pdf generation

+++ run-all-tests.sh
@@ -1,4 +1,8 @@
-pip uninstall pipelineutilities
+if [ -z "$1" ]
+  then
+    pip uninstall pipelineutilities
 
-./scripts/codebuild/install.sh  ||  { exit 1; }
+    ./scripts/codebuild/install.sh  ||  { exit 1; }
+fi
+
 ./scripts/codebuild/pre_build.sh ||  { exit 1; }
